Rename cargo_intake package to cargo
