+++ src/main/java/org/team1540/robot2019/OI.java
@@ -8,9 +8,9 @@ import edu.wpi.first.wpilibj.buttons.Button;
 import edu.wpi.first.wpilibj.buttons.JoystickButton;
 import edu.wpi.first.wpilibj.command.Command;
 import org.apache.log4j.Logger;
-import org.team1540.robot2019.commands.cargo_intake.EjectThenDown;
-import org.team1540.robot2019.commands.cargo_intake.LoadingStationCargoIntake;
-import org.team1540.robot2019.commands.cargo_intake.LowerThenCargoIntake;
+import org.team1540.robot2019.commands.cargo.EjectThenDown;
+import org.team1540.robot2019.commands.cargo.LoadingStationCargoIntake;
+import org.team1540.robot2019.commands.cargo.LowerThenCargoIntake;
 import org.team1540.robot2019.commands.climber.ClimbLevelThreeGyro;
 import org.team1540.robot2019.commands.climber.ClimbLevelTwo;
 import org.team1540.robot2019.commands.climber.PrepareForClimb;
