+++ src/main/java/org/team1540/robot2019/commands/cargo/EjectCargo.java
@@ -1,4 +1,4 @@
-package org.team1540.robot2019.commands.cargo_intake;
+package org.team1540.robot2019.commands.cargo;
 
 import edu.wpi.first.wpilibj.command.TimedCommand;
 import org.apache.log4j.Logger;
