+++ src/main/java/org/team1540/robot2019/commands/cargo/LowerThenCargoIntake.java
@@ -1,4 +1,4 @@
-package org.team1540.robot2019.commands.cargo_intake;
+package org.team1540.robot2019.commands.cargo;
 
 import edu.wpi.first.wpilibj.command.CommandGroup;
 import org.team1540.robot2019.Robot;
